Added views generator and also removed minor bugs for other controllers

+++ src/Generators/ControllerGenerator.php
@@ -39,7 +39,7 @@ class ControllerGenerator
         ?>
 
 _EOT;
-      file_exists(!$this->name.'.php');
+      if(!file_exists($this->name.'.php'))
       {
         file_put_contents($this->controller_a_path.'/'.$this->name.'.php',$str);
       }
